Support devel 'unknown' pelican version
When using a development version of pelican, the version is set to
'unknown'. However, this cannot be parsed as an integer, and results in
failure:

	ValueError: invalid literal for int() with base 10: 'unknown'

The proposed fix is to consider 'unknown' version as a >=4 version. The
global variable has been moved into the if statement instead, so that
lazy evaluation avoids the cast to an integer if the version is not
known.